Controllers Updates-1

+++ internal/api/product/product_controller.go
@@ -14,6 +14,8 @@ type ControllerProduct struct {
 	categoryService *category.ServiceCategory
 }
 
+// @BasePath /product
+
 func NewProductController(service *product.ServiceProduct, service2 *category.ServiceCategory) *ControllerProduct {
 	return &ControllerProduct{
 		productService:  service,
@@ -21,6 +23,16 @@ func NewProductController(service *product.ServiceProduct, service2 *category.Se
 	}
 }
 
+// GetAll godoc
+// @Summary Gets all products with pagination parameters page and size
+// @Tags Product
+// @Accept  json
+// @Produce  json
+// @Param page query int false "Page Index"
+// @Param size query int false "Page Size"
+// @Success 200 {object} pagination.Pages
+// @Failure 404 {object} map[string]string
+// @Router /product [get]
 func (c *ControllerProduct) GetAll(g *gin.Context) {
 	pageIndex, pageSize := pagination.GetPaginationParametersFromRequest(g)
 	products, count := c.productService.GetAll(pageIndex, pageSize)
@@ -39,7 +51,19 @@ func (c *ControllerProduct) GetAll(g *gin.Context) {
 	g.JSON(http.StatusOK, paginatedResult)
 }
 
-func (c *ControllerProduct) CreateProduct(g *gin.Context) {
+// Create godoc
+// @Summary Creates products with the given request
+// @Tags Product
+// @Accept  json
+// @Produce  json
+// @Param product-request body ResponseProduct true "Takes the products and adds them to db"
+// @Success 201 {object} map[string]string
+// @Failure 400 {object} map[string]string
+// @Failure 500 {object} map[string]string
+// @Security ApiKeyAuth
+// @param Authorization header string true "Authorization"
+// @Router /product [post]
+func (c *ControllerProduct) Create(g *gin.Context) {
 	var req product.Product
 	err := g.ShouldBind(&req)
 	if err != nil {
@@ -49,9 +73,11 @@ func (c *ControllerProduct) CreateProduct(g *gin.Context) {
 		g.Abort()
 		return
 	}
+	//there is no category id it will be default category named "other"
 	if req.CategoryID == 0 {
 		req.CategoryID = c.categoryService.GetCategoryWithCode(0).ID
 	}
+	//checking given request category id if is not exist returns
 	category := c.categoryService.GetCategoryWithId(int(req.CategoryID))
 	if category.ID == 0 {
 		g.JSON(http.StatusBadRequest, gin.H{
@@ -60,6 +86,7 @@ func (c *ControllerProduct) CreateProduct(g *gin.Context) {
 		g.Abort()
 		return
 	}
+	//creating product
 	productItem := product.NewProduct(req.Name, req.Price, req.Amount, req.Code, req.Description, req.CategoryID)
 	err = c.productService.Create(productItem)
 	if err != nil {
@@ -75,11 +102,25 @@ func (c *ControllerProduct) CreateProduct(g *gin.Context) {
 	})
 }
 
-func (c *ControllerProduct) SearchProduct(g *gin.Context) {
+// Search godoc
+// @Summary Gets all products with search parameters name, amount or categoryId
+// @Tags Product
+// @Accept  json
+// @Produce  json
+// @Param name query string false "Name"
+// @Param categoryId query int false "CategoryId"
+// @Param amount query int false "Amount"
+// @Success 200 {object} ResponseProduct
+// @Failure 400 {object} map[string]string
+// @Failure 404 {object} map[string]string
+// @Router /product/search [get]
+//Search works with query parameters name, category id and amount, amount process looks equal or if there is more
+func (c *ControllerProduct) Search(g *gin.Context) {
 	var products []product.Product
 	var req product.Product
 	var err error
 
+	//taking query parameters
 	name, isOk := g.GetQuery("name")
 	categoryId, isOk1 := g.GetQuery("categoryId")
 	amount, isOk2 := g.GetQuery("amount")
@@ -88,6 +129,7 @@ func (c *ControllerProduct) SearchProduct(g *gin.Context) {
 		req.Name = name
 	}
 	if isOk1 {
+		//checking category-id query parameter
 		intCategory, err := strconv.Atoi(categoryId)
 		if err != nil {
 			g.JSON(http.StatusBadRequest, gin.H{
@@ -106,6 +148,7 @@ func (c *ControllerProduct) SearchProduct(g *gin.Context) {
 		req.CategoryID = uint(intCategory)
 	}
 	if isOk2 {
+		//checking amount query parameter
 		intAmount, err := strconv.Atoi(amount)
 		if err != nil {
 			g.JSON(http.StatusBadRequest, gin.H{
@@ -123,7 +166,7 @@ func (c *ControllerProduct) SearchProduct(g *gin.Context) {
 		}
 		req.Amount = uint(intAmount)
 	}
-
+	//if query parameters exist
 	if isOk || isOk1 || isOk2 {
 		products, err = c.productService.Search(&req)
 		if err != nil {
@@ -134,19 +177,47 @@ func (c *ControllerProduct) SearchProduct(g *gin.Context) {
 			return
 		}
 	}
-
+	//if there is no product
 	if len(products) == 0 {
-		g.JSON(http.StatusBadRequest, gin.H{
+		g.JSON(http.StatusNotFound, gin.H{
 			"error_message": "No data found with search parameters.",
 		})
 		g.Abort()
 		return
 	}
+	//shaping response products
+	var response []ResponseProduct
+	for _, p := range products {
+		var resProduct ResponseProduct
+		resProduct.ID = p.ID
+		resProduct.Name = p.Name
+		resProduct.Amount = p.Amount
+		resProduct.Code = p.Code
+		resProduct.Price = p.Price
+		resProduct.Description = p.Description
+		resProduct.CategoryId = p.CategoryID
+		response = append(response, resProduct)
+	}
 
-	g.JSON(http.StatusOK, products)
+	g.JSON(http.StatusOK, response)
 }
 
-func (c *ControllerProduct) DeleteProduct(g *gin.Context) {
+// Delete godoc
+// @Summary Deletes given product id, it can be with query or request body
+// @Tags Product
+// @Accept  json
+// @Produce  json
+// @Param id query int false "Takes the product id and deletes it, product id is needed"
+// @Param product-request body ResponseProduct false "Takes the product infos and deletes it, product id is needed"
+// @Success 200 {object} map[string]string
+// @Failure 400 {object} map[string]string
+// @Failure 404 {object} map[string]string
+// @Security ApiKeyAuth
+// @param Authorization header string true "Authorization"
+// @Router /product [delete]
+//Delete works with query parameter product id and also works with product request body it should contain exist product id
+func (c *ControllerProduct) Delete(g *gin.Context) {
+	//checking delete from query
 	id, isOK := g.GetQuery("id")
 	if isOK {
 		deg, err := strconv.Atoi(id)
@@ -160,13 +231,14 @@ func (c *ControllerProduct) DeleteProduct(g *gin.Context) {
 
 		err = c.productService.DeleteWithID(deg)
 		if err != nil {
-			g.JSON(http.StatusBadRequest, gin.H{
+			g.JSON(http.StatusNotFound, gin.H{
 				"error_message": err.Error(),
 			})
 			g.Abort()
 			return
 		}
 	} else {
+		//checking delete from request body
 		var req product.Product
 		err := g.ShouldBind(&req)
 		if err != nil {
@@ -176,9 +248,16 @@ func (c *ControllerProduct) DeleteProduct(g *gin.Context) {
 			g.Abort()
 			return
 		}
+		if req.ID == 0 {
+			g.JSON(http.StatusBadRequest, gin.H{
+				"error_message": "Check your request body. Id is need",
+			})
+			g.Abort()
+			return
+		}
 		err = c.productService.Delete(&req)
 		if err != nil {
-			g.JSON(http.StatusBadRequest, gin.H{
+			g.JSON(http.StatusNotFound, gin.H{
 				"error_message": err.Error(),
 			})
 			g.Abort()
@@ -186,13 +265,25 @@ func (c *ControllerProduct) DeleteProduct(g *gin.Context) {
 		}
 	}
 
-	g.JSON(http.StatusAccepted, gin.H{
+	g.JSON(http.StatusOK, gin.H{
 		"message": "Product successfully deleted.",
 	})
 
 }
 
-func (c *ControllerProduct) UpdateProduct(g *gin.Context) {
+// Update godoc
+// @Summary Updates given product
+// @Tags Product
+// @Accept  json
+// @Produce  json
+// @Param product-request body ResponseProduct true "Takes the product infos and updates it. Product id, name, category id and code needed"
+// @Success 200 {object} map[string]string
+// @Failure 400 {object} map[string]string
+// @Security ApiKeyAuth
+// @param Authorization header string true "Authorization"
+// @Router /product [put]
+//Update works with all product fields and changes on it
+func (c *ControllerProduct) Update(g *gin.Context) {
 	var req product.Product
 	err := g.ShouldBind(&req)
 	if err != nil {
